refactor(guard): isolate WebAuthn transport adapters
